Incremental crawl

+++ cli/src/main/java/com.datagenio.cli/CrawlRunner.java
@@ -1,16 +1,17 @@
 package com.datagenio.cli;
 
-import com.datagenio.crawler.SimpleCrawler;
+import com.datagenio.crawler.PersistentCrawler;
 import com.datagenio.context.Context;
 import com.datagenio.crawler.browser.BrowserFactory;
 import com.datagenio.databank.InputBuilderFactory;
 import com.datagenio.generator.GeneratorImpl;
+import com.datagenio.generator.api.Generator;
 import com.datagenio.generator.converter.GraphConverterImpl;
 import com.datagenio.generator.converter.BodyConverter;
 import com.datagenio.generator.converter.HttpRequestAbstractor;
 import com.datagenio.generator.converter.StateConverter;
 import com.datagenio.generator.converter.UrlAbstractor;
-import com.datagenio.model.api.WebFlowGraph;
+import com.datagenio.model.WebFlowGraph;
 import com.datagenio.storage.Neo4JReadAdapter;
 import com.datagenio.storage.Neo4JWriteAdapter;
 import com.datagenio.storage.connection.ConnectionResolver;
@@ -84,44 +85,52 @@ public class CrawlRunner {
 
         // Validate output dir
         String directory = arguments.getOptionValue(ArgumentParser.OUTPUT);
-        if (!outputDirIsValid(directory)) {
+        if (!outputDirIsValid(directory, !continuePrevious(arguments))) {
             System.out.println("Please check directory exists and can be written to.");
             return;
         }
 
         System.out.println("Preparing dependencies...");
+        Generator generator = getGenerator(arguments);
 
+        // Begin modeling site
+        System.out.println("Beginning modeling process...");
+        WebFlowGraph model = generator.generateWebModel();
+
+        System.out.println("Finished modeling site.");
+        System.out.println("Found " + model.getStates().size() + " states, and " + model.getTransitions().size() + " transitions.");
+
+        System.out.println("Generating data set...");
+        generator.generateDataset(model);
+        System.out.println("Finished writing data set.");
+    }
+
+    private static GeneratorImpl getGenerator(CommandLine arguments) {
         var jsonBuilder = new GsonBuilder().setPrettyPrinting();
         var gson = jsonBuilder.create();
 
         var context = getContext(arguments);
-        var crawler = new SimpleCrawler(context, BrowserFactory.drivenByFirefox(), InputBuilderFactory.get());
-        var readAdapter = new Neo4JReadAdapter(context.getConfiguration());
-        var writeAdapter = new Neo4JWriteAdapter(context.getConfiguration(), ConnectionResolver.get(context.getConfiguration()), gson);
+        var connection = ConnectionResolver.get(context.getConfiguration());
+        var readAdapter = new Neo4JReadAdapter(context.getConfiguration(), connection);
+        var writeAdapter = new Neo4JWriteAdapter(context.getConfiguration(), connection, gson);
         var urlAbstractor = new UrlAbstractor();
         var bodyConverter = new BodyConverter();
         var requestAbstractor = new HttpRequestAbstractor(urlAbstractor, bodyConverter);
         var stateConverter = new StateConverter(urlAbstractor, requestAbstractor, context.getRootUri());
 
-        // Begin modeling site
-        System.out.println("Beginning modeling process...");
+        context.setReadAdapter(readAdapter);
+        context.setWriteAdapter(writeAdapter);
 
-        var generator = new GeneratorImpl(context, crawler, new GraphConverterImpl(context, stateConverter, requestAbstractor), readAdapter, writeAdapter);
-//        EventFlowGraph graph = generator.crawlSite();
-        WebFlowGraph model = generator.generateWebModel();
-
-//        System.out.println("Finished crawling site.");
-//        System.out.println("Found " + graph.getStates().size() + " states, and " + graph.getTransitions().size() + " transitions.");
-
-        System.out.println("Finished modeling site.");
-        System.out.println("Found " + model.getStates().size() + " states, and " + model.getTransitions().size() + " transitions.");
-
-        generator.generateDataset(model);
-        System.out.println("Finished writing dataset.");
+        var crawler = new PersistentCrawler(context, BrowserFactory.drivenByFirefox(), InputBuilderFactory.get());
+        return new GeneratorImpl(context, crawler, new GraphConverterImpl(context, stateConverter, requestAbstractor), readAdapter, writeAdapter);
     }
 
     private static boolean isVerbose(CommandLine arguments) {
         return arguments.hasOption(ArgumentParser.VERBOSE);
+    }
+
+    private static boolean continuePrevious(CommandLine arguments) {
+        return arguments.hasOption(ArgumentParser.CONTINUE);
     }
 
     private static int getMaxDepth(CommandLine arguments) {
@@ -142,6 +151,8 @@ public class CrawlRunner {
                 getMaxDepth(arguments)
         );
 
+        context.setContinueExistingModel(continuePrevious(arguments));
+
         System.out.println("Max exploration depth: " + context.getCrawlDepth());
         return context;
     }
@@ -153,7 +164,7 @@ public class CrawlRunner {
         return validator.isValid(url);
     }
 
-    public static boolean outputDirIsValid(String dir) {
+    public static boolean outputDirIsValid(String dir, boolean clear) {
         if (dir == null) {
             return false;
         }
@@ -163,7 +174,7 @@ public class CrawlRunner {
             directory.mkdir();
         }
 
-        if (directory.list().length > 0) {
+        if (clear && directory.list().length > 0) {
             try {
                 FileUtils.deleteDirectory(directory);
                 directory.mkdir();
